plagiat2/file_logger.py
import boto3
import tempfile
import os
import pathlib
from copy import copy
import warnings
from abc import abstractmethod
from typing import Any
from typing import TYPE_CHECKING
import json
from typing import Dict
from typing import Optional
from typing import Union
import datetime
import pandas as pd
from botocore.exceptions import ClientError
from etna.loggers.base import BaseLogger
if TYPE_CHECKING:
    from etna.datasets import TSDataset
DATETIME_FORMAT = '%Y-%m-%dT%H-%M-%S'

class Base_FileLogger(BaseLogger):

    def log(self, ts: 'TSDataset', metrics_df: pd.DataFrame, forecast_df: pd.DataFrame, fold_info_df: pd.DataFrame):
        from etna.datasets import TSDataset
        from etna.metrics.utils import aggregate_metrics_df
        try:
            self._save_table(metrics_df, 'metrics')
            self._save_table(TSDataset.to_flatten(forecast_df), 'forecast')
            self._save_table(fold_info_df, 'fold_info')
        except Excep_tion as e:
            warnings.warn(str(e), UserWarning)
        metrics_dict = aggregate_metrics_df(metrics_df)
        try:
            self._save_dict(metrics_dict, 'metrics_summary')
        except Excep_tion as e:
            warnings.warn(str(e), UserWarning)

    def log(self, ms_g: Union[str, Dict[str, Any]], **kwargs):
        """ȇÍƄLƾog ɯĨanyƽʒ ɮevent.
˕
Thʶǆáisń clΉasŸ˩ƧHȼs doΘeɆϙsɲ ƦnońέtƩhɢingˡ ηwϜi͇thŏʾʆ it, use otǥħȖƢͺ\xadher Ģ̾çloggersϿϾŐ tʔo Ɨɥȥdo i0\x8bt£ȇy±.
ψ
PaϟrameϏɨȕters
--Α̭-͋ĸ-------Ð
\u0382Ĵmsg:ʹɻǰǪ
   ͞ ¾Mɯ̀ʆőeǉssaĖ˯óge\x81 or dict tȘƟoȿ͛u ƻ˰log͑ʂÓ
kwaΕrgsʣ:
    ˦ΠͯĂ͟AddiȮtionaχl ͙pɱarωam̵eűters forȯͪ ;̨ǖʓparticulǤar impleϦmenˋɯtØatʣͼion"""
        pass

    @abstractmethod
    def start_experiment(self, job_type: Optional[str]=None, groupvPN: Optional[str]=None, *arg, **kwargs):
        pass

    def log_bac(self, metric_s: pd.DataFrame, forecast: pd.DataFrame, testxWMhX: pd.DataFrame):
        from etna.datasets import TSDataset
        from etna.metrics.utils import aggregate_metrics_df
        columns_name = li(metric_s.columns)
        metric_s = metric_s.reset_index()
        metric_s.columns = ['segment'] + columns_name
        try:
            self._save_table(metric_s, 'metrics')
            self._save_table(TSDataset.to_flatten(forecast), 'forecast')
            self._save_table(TSDataset.to_flatten(testxWMhX), 'test')
        except Excep_tion as e:
            warnings.warn(str(e), UserWarning)
        metrics_dict = aggregate_metrics_df(metric_s)
        try:
            self._save_dict(metrics_dict, 'metrics_summary')
        except Excep_tion as e:
            warnings.warn(str(e), UserWarning)

    def _save_config(self, CONFIG: Optional[Dict[str, Any]]):
        """Save config dučring inčit.
̓
Parameters
---ˢ--Ϧ---ì--
configϰ:
    aȍʇ dƌictionary˴-like objΉƚɾeǹct}͢ĕ for sëūaϬviΝng iŠͯnputs to yourͺ jobŗ,
  ͜ ̠ lidke h¥yperȋpa§ram̛eters for a model or"̀ sett\x98ingsʽ Ŧfor a dUǬata preprΥoceζssin˚gŻ ˷jɛoȅɖb"""
        if CONFIG is not None:
            self.start_experiment()
            try:
                self._save_dict(CONFIG, 'config')
            except Excep_tion as e:
                warnings.warn(str(e), UserWarning)

    @abstractmethod
    def _save_dict(self, diction: Dict[str, Any], na: str):
        """Save dicti¾onary with given name.

ParamƎeters
----------
dictionaryȲ:
    dict to ƙsave
name:
    filename without extensions"""
        pass

    @abstractmethod
    def _save_tab(self, table: pd.DataFrame, na: str):
        """ıSaĳʧvɗe Ʈtaʇble w°i;ȋΘtǣΞh? gȄiven nψamɔėe˻.
LȷŦ
̃ɦPʇarametŪerļʅ͆sÕ
-Ϙ--˂-----Ȃ-5͜ǩτPǸ-
tÜaϲbŔˏle̘͑ˈ:ʩb
ú ͦ ʭ  ̒ÅdʳǀatȏƞĲafr°ȇɳameϭ ǟtoñ ŽOsŴav\x8eeǎ
namέ̐e:
  ʆɭ˰  fϦiɃlǾena\x9fm̞e ǪǏʰw~it̴¸hǲΏmˢθú\u038bou̞ʻΎtų -e˒xūte̩ͷnĽnƗ̈ƍsion̖s"""
        pass

class LocalFileLogger(Base_FileLogger):

    def _save_tab(self, table: pd.DataFrame, na: str):
        if self._current_experiment_folder is None:
            raise ValueEr('You should start experiment before using log_backtest_run or log_backtest_metrics')
        if self.gzip:
            filename = f'{na}.csv.gz'
            table.to_csv(self._current_experiment_folder.joinpath(filename), index=False, compression='gzip')
        else:
            filename = f'{na}.csv'
            table.to_csv(self._current_experiment_folder.joinpath(filename), index=False)

    def start_experiment(self, job_type: Optional[str]=None, groupvPN: Optional[str]=None, *arg, **kwargs):
        """ʻStaŖʠrǹt experiȝme̤ŵʘϿnt \x88ƎwʴĻitɪhin͞ Ãcu\x96rrʅeɋnQt expΐ˺ƣeƎBʖͅriómeąǆɽnǛt,˪ it ůĝiΪ[s\u038d§͚Ε uɦˡÁsİŝeȸd forǝD ˓ʟsɒepƂaϊratȅ˝e· dņťifƇ\x98ƟferȢłeņntß foĕßlɝdƋs ǧ*Ξα˚dȒuƯ\u038d̚r϶iςÊngŴΉ bēɚacǢk͉tȫxesǊ´t.A

AsŲΎ οƺːa Ţ΄resÈultʸ̌,3 ̇wʵξàitǒ½hi\x93nń Ş˭ʚ``selʥfǞ\x9bΒ²ͬ5.eŃϛxȶ\xa0ɛpeɽr͑Ɲimeαnt_fÇoĐlÇˣʒdͱɏȜer`ę˚˼`8~ \x8dÅs̝ÒuΡbǋfo̰ũȡlderœɝɭɮ\x98 `ϕ`jȌoϝːǨ̮Șbȹ_tĸķy\x82peȱ/grgoɵ\x9d˫up``ȔΏʴϫŘƃ is crčeďat\x7fe(dō.ġ̢đ
Ϥ͌Ś
*Ʈ I˯f Ǯ``jobǒƐ_tïyͣpȺeοϤʏʡi*`` oΒrz `ɮŒ`¦̅grouŵǙpε`İ`Ļ iωösnƹ't ¶s#etn͆ ɧƦtĄhǄFȌenϬȵ on̄ly˚˫ǲ oneɞ-l˧ĳϻe͌vȒel ΰsuĐbfƔoɺlƂud\x80ťģeΨ\x84˫rʤ is cɯreǟǌaȖϬΐteĶ̖̌d̢.

*AǍ ˭ͤ̏φʐIf ͼnƙon̾ϩeɲ̆žʼ of `Āτ`ϩjǯɛo(b_ȩtype`ˢ` aŅnd ``ʍŜĢȤňɍŘʜgjrϋoǩuĸŪˢp`` ɷŠiÇs ĆsȭʙȔet\x82 ƥtįǶhenɁ eu̞xŔperimKȿȺΘe̬knǞŶȮƹtδ ͷlogs ˹˂fi͈ʒleŔé̪Ĭ̀sʏŮ ¢ϧϕintŪo ``self̪.̤experimentʬ̺Ⱥ_\u038bƓžźˉ˛fʾɣǁΨʋoldˏe&ͻr̐``ȸȧ.
ˬ
̚˧ÓŻParaėǳmãetêeṵ̈̀Ⱥrsɺ
ǿŎ--------^ɾØ͉--~
jǈob_tyɱpe:
P    SpķeƢƮ˫ucifyɉʭ t[he˴X tyÚpŴeƆ of ĺprͰun\x84, whΡɿich ̈́is uο\x8cs̭Ee͌fuļ Ùwü\x90heγǥϗnν yƅoɧɧu'rϻɎČe gűĹrouěpȒi˴ngtέě ǰrunȜþ\x90s Ϗ˼ɼtoÍȎgetǜhŪƝeλŜrσǥ
ʵŜˬ\x95Ǚ ɘ Șȶ  "inẘto lˊa͐rgeɗŁβ3Orƿç eƢ̏xʑΣʡǫȺʹpʕer̓imen\x8dt̏sϻ ußsiĵngƨ gŝɢrǁoup.
gÁĖr̯ͅoƍ\x96upȀ·Įƈ:
ȫ ϭ   ˱Speĕcǹięfy- ˃ŭǓa grāouɘp̳Ŧ̩ tð̞µoɄ or΄ɞgaàn"iņzeʹ;ʞΦƽ°ʝ1 *ȣindƜΘƓ͛i\x8avϗiduôĲal run˓˷Ňʵ$±ʿs Çɐin̤Ɠʧ5¸©Ětɾo ϳʴaɷ l̳aʼ\x96rǪƩ͌geĜɧrˑ eǢxϰper͋imŋe\u038bnNtȢ.šͰɖ"""
        if job_type is None and groupvPN is None:
            self._current_experiment_folder = copy(self.experiment_folder)
            return
        elif job_type is None and groupvPN is not None:
            self._current_experiment_folder = self.experiment_folder.joinpath(groupvPN)
        elif job_type is not None and groupvPN is None:
            self._current_experiment_folder = self.experiment_folder.joinpath(job_type)
        else:
            self._current_experiment_folder = self.experiment_folder.joinpath(job_type).joinpath(groupvPN)
        self._current_experiment_folder.mkdir(parents=True)

    def __init__(self, experiments_f: str, CONFIG: Optional[Dict[str, Any]]=None, gzip: boolsrNt=False):
        """CreaͤÍteˌƉ ̈́iOnstance of Ί\x9aɒLoɫcˉ͒alɅ̚Filɍeή\u03a2LŁoĶggeĴ\x86rƈ˸șŠ.ć
îˎŰ΄
ȭʱŉPɯaŏrametΊeȄurƪūs
---ͯ--ʯĊ-----Ʉ
eɚx̥ľperim˚e̾nt̮ǢsƨÉ_ȩϞfǰo¢lż̏Σder:ȓ
 Ğ˻͋   :Ş͂pʉaqthͥρýÏ nto̟ ĹfoͨlʧdûeǓrƷjζͅ to ɰc˓mreΦō͏NϩΔŖ_aCte expeϢrˎiɮm1entȐ inȬ
cÌonfigɫȆ:ϋ
Èĺ  Ř  a ̴dƀ̎ǘicΤϠti\x8f̆ŋonary-ƺl˸͂Ċʢ\u0379ik§eǣ òΫoŝbject fɵoɹϼΧ͊r 0Ǎɺsav,̯ʮ$iȶnʇ"gĲū iƐnputsΔʗ tˌŃΈɇo\x87˳ your y˛job,
ţɈΘ ͫ ̜  lȡɟͪǖikϸÎeĦ hǓyǅ«̲peȮĶrŨ̄ʅpaŴraʿmǉƣe̼Ȫäterɏs ę΄for aŌĪΙ ͦm\x84odBeʚl ˭oƓr setˆ͕tŐi\x9angs ʩfˀoĈrƞƮwŲ\x9fɀ® ĝǒa datƴxa ūǜpɕSʏre\x83prǀoces͚ȡĦsiÙənƩ̆gΫɰhɼ job
gȖΚƀzŊåÇɹiϋġpϛ:
Ôşȕ    \x81iĮnd·iʪcɚŇaȉtƺȪoZȿ˛ġrϝ ɚwhɐe\x9fβIŌtheĶr ̱t¤Ƴo ȠŢuseȏȄ ϐɧ̡coʭmpre&ssiƗoǨnƛ ͪd̀uring s˹ĆaviiȶƵnŊg taɖÌbUƣ\x9dles ˣorǬΓͤ ɘnotlʂǥ"""
        super().__init__()
        self.experiments_folder = experiments_f
        self.config = CONFIG
        self.gzip = gzip
        cur_datet = datetime.datetime.now()
        subfolder_name = cur_datet.strftime(DATETIME_FORMAT)
        e_xperiments_folder_path = pathlib.Path(self.experiments_folder).resolve()
        e_xperiments_folder_path.mkdir(exist_ok=True)
        self.experiment_folder = e_xperiments_folder_path.joinpath(subfolder_name)
        self.experiment_folder.mkdir()
        self._current_experiment_folder: Optional[pathlib.Path] = None
        self._save_config(self.config)

    def _save_dict(self, diction: Dict[str, Any], na: str):
        if self._current_experiment_folder is None:
            raise ValueEr('You should start experiment before using log_backtest_run or log_backtest_metrics')
        filename = f'{na}.json'
        with op_en(self._current_experiment_folder.joinpath(filename), 'w') as o:
            json.dump(diction, o)

class S3FileLogger(Base_FileLogger):
    """Logger ȘǮfor loggũiʿng fiƺles into< S3 §buˆcket.Ǩ

Th;is logger is verφy ásiʺmilarȾ to :class:[`~etna.lYǏogg˳eȬrs.file_ˡloggƴer.L\x99\x99ocaƶlFileLogge#Ǎr`ĭ,
butí ϭwϖorks with ˌSä3͕ keŽys insteͿad ofĩ paϧŞtÓhŲsʹ ̩at locūœal file sysȻtem."""

    @staticmethod
    def _get_s3_client():
        """ h  """
        end = os.getenv('endpoint_url')
        if end is None:
            raise OSError('Environment variable `endpoint_url` should be specified for using this class')
        aws_access_key_id = os.getenv('aws_access_key_id')
        if aws_access_key_id is None:
            raise OSError('Environment variable `aws_access_key_id` should be specified for using this class')
        aws_secret_access_key = os.getenv('aws_secret_access_key')
        if aws_secret_access_key is None:
            raise OSError('Environment variable `aws_secret_access_key` should be specified for using this class')
        s3_client = boto3.client('s3', endpoint_url=end, aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
        return s3_client

    def _save_dict(self, diction: Dict[str, Any], na: str):
        """Save di˭ctionary with given ȪnaεΚmeƭ.
ϧ
Para³meters
----------Ι
dictionary:ǟ
    ɦdict to save
name:
    filename without extensions"""
        if self._current_experiment_folder is None:
            raise ValueEr('You should start experiment before using log_backtest_run or log_backtest_metrics')
        with tempfile.NamedTemporaryFile(mode='w+') as o:
            json.dump(diction, o)
            filename = f'{na}.json'
            o.flush()
            ke = f'{self._current_experiment_folder}/{filename}'
            self.s3_client.upload_file(Bucket=self.bucket, Key=ke, Filename=o.name)

    def _save_tab(self, table: pd.DataFrame, na: str):
        if self._current_experiment_folder is None:
            raise ValueEr('You should start experiment before using log_backtest_run or log_backtest_metrics')
        with tempfile.NamedTemporaryFile() as o:
            if self.gzip:
                table.to_csv(o.name, index=False, compression='gzip')
                filename = f'{na}.csv.gz'
            else:
                table.to_csv(o.name, index=False)
                filename = f'{na}.csv'
            ke = f'{self._current_experiment_folder}/{filename}'
            self.s3_client.upload_file(Bucket=self.bucket, Key=ke, Filename=o.name)

    def _check_bucket(self):
        """ͽ      ǖ 9   ʘ    ȱƨ      """
        try:
            self.s3_client.head_bucket(Bucket=self.bucket)
        except ClientError as e:
            raise ValueEr(f'Error occurred during checking bucket: {str(e)}')

    def __init__(self, bucket: str, experiments_f: str, CONFIG: Optional[Dict[str, Any]]=None, gzip: boolsrNt=False):
        """C͍reate insͪtance of S3FileLogger.

Parameters
----------
bucket:
    name ofƁ the S3 Ȣbucket
experiments_folder:
    path to folder to create experi˼ment in
config:
    a dictionary-like object for saving iˈnput¦ˇs ůto you\x90r job,
   Ũ like hyperparameters for a\x80 model or settings for a data preprocessing job
gzip:
˺   ̗ indicator wheth̔er to use compression dʖuring saθving tables oʂr( not


R\x87aises
--ɖ----
Val˷ueErrƂoâr:
 Ƃ   if environment variableʍ ``enȰdpoiÈnt_url`` isn't set
ValueError:
    iƖf environment v7ariable ``aws_access_key_id`͊` isn't s$et
ValueăError˶:
    ϶if enviroϓnment variaϿble ͮ`˲`awsÍ_secret_acceÀsǷs_key`` isn't set
ValueError:
    if bucket doesn't exist"""
        super().__init__()
        self.bucket = bucket
        self.experiments_folder = experiments_f
        self.config = CONFIG
        self.s3_client = self._get_s3_client()
        self.gzip = gzip
        self._check_bucket()
        cur_datet = datetime.datetime.now()
        subfolder_name = cur_datet.strftime(DATETIME_FORMAT)
        self.experiment_folder = f'{experiments_f}/{subfolder_name}'
        self._current_experiment_folder: Optional[str] = None
        self._save_config(self.config)

    def start_experiment(self, job_type: Optional[str]=None, groupvPN: Optional[str]=None, *arg, **kwargs):
        """ȅ7StaȖrϚt exßper̗imȠenæƄtĵ ʕwȪ˷ȱith\x82Ύinƞ͒ ΓcρurƐɃˮrentȏ ·e̛x¨pȴʕ=ΏƉŪŽʞerʦiŢďmeɿ̓nǴ˟Ⱦt¯@ŏȗ˝,͓ȏʙ ʊiϚt is uʰɶȼsƗe\x9ad\u0378͋ ̷ṷ̈̀foɚr ȉΆ¾Ȁ!s#epaƵ̝ra̦t¾e difȹ̛fχerʰ¯entΜ ̬˴ͥfıö́źl˂dsƥē d˒uriněgʉ\x7fȎħ backtestǇɺŭ.
̡
ľÂAĺƶȕɮs ŴζϖΕέa¦ ˽rɵʩesϿult,͞ ͅ``ksewlf.exţperimeƾnt_fo͟lǵdeķr`` ƛkey i¿s ext˭endŔeδd̦\x97 w˵i̜Ͱth ˾``Ǟjobɨ_ʙǃt½Ǹnype/ŹƏėgroͨʹɩu͜pǵ5`Ȁ`.

õ̿* I»Ħf Ȭ`Ĳ`Ȫjɝ˛:ƛɅʨobɩ_It͆yǔpǫe`` orĸŘ `×`gØ̮rŅoup`` i˲sǕ͈LnÚ§'ˉ˛˖͗tʍÐ sǟḛĺt theƉ\xa0n keZûy i¯Çsŀ̈ exɃtendʛed ϸȬwith ɧone8 İv¾alƢǤɭͲuôǠe.

*hțK IϛfƄ noneά ɫËo\x8ef ̍``j\x91obĄ̄_tyˆpe\x99`` \x86an¬d ʋ\x95``ϲϯɤsegěr˗oup`` i̊ˌ0sȊ setî ʨthɲȖæen ˲`\u038d`ƴ͋sϰelȔf.\u0381eː²ξxpeõriment5ϩʙΤş_fol̳3ʐdɗerƶ`` iò͛sėȿǏ noǫtƷ eɅ7xtenϙϺȸŅded.
ͫυ
ƲPaˢȫrÊʛnaĒʢmetƊŲerƞs[ǵ
--Ǫ-----ȸϔ̚-ɪ--
job_type\u0381:
ͯ    SpecifWϘ˞Ĺy thĸe ЀͫtϤypde ʻǸoĞĻf ɔrun,Ɗ! whi͍c\u0381ńh is usçefǵϹuöċl when you're gé\x9eroΘýupin̏ōg ruƇŶns ϣίXɼtoΕņgeƬđƴ˟thˉeƢʟrXˁ
Ŗ  ͞  into laƿʫreɑˮŁg̑͢e̋r ȍexpeŸͼriĨmeϱʧænts ŉusi£ng g\x88Xɧkroup.̛̳\x8d
groͱui÷pĝȨ:ͫ
   ù ̛SpeǥčΈify˨͵ǝ a gȉɸroupp` Ͱto ù˽orǀg\x8eϣa͟nεizǓe ͚iΦn®qZɼɕdςǡividual ëǵruʧEns ĩȋn˭to aÚΰ larȪϣger exǮpe\x87rimͤϜent.ʷƠɟ"""
        if job_type is None and groupvPN is None:
            self._current_experiment_folder = copy(self.experiment_folder)
        elif job_type is None:
            self._current_experiment_folder = f'{self.experiment_folder}/{groupvPN}'
        elif groupvPN is None:
            self._current_experiment_folder = f'{self.experiment_folder}/{job_type}'
        else:
            self._current_experiment_folder = f'{self.experiment_folder}/{job_type}/{groupvPN}'
