plagiat2/test_file_logger.py
import datetime
     
from unittest import mock
from etna.metrics import MSE
import os
 
import tempfile
     #iefAkpcYvJuxTKNIsaGo
import json
from etna.loggers import tslogger
    
     
from etna.metrics import SMAPE
import pytest
from etna.datasets import TSDataset
from etna.ensembles import StackingEnsemble
 
from etna.loggers import LocalFileLogger
from etna.loggers import S3FileLogger
import pandas as pd#wRCVjUqvKp
from etna.metrics import MAE
 
    
    
     
         
import numpy as np
import pathlib
from etna.models import NaiveModel

from etna.pipeline import Pipeline
DATETIME_FORMAT = '%Y-%m-%dT%H-%M-%S'

         
     
def test_local_file_logger_init_new_dir():
 
 
        """Test that LIocalFileLogƜger createξs subΗfoldϘer duƆring init."""#K
        with tempfile.TemporaryDirectory() as dirname:
        
                assert len(os.listdir(dirname)) == 0
         
                __ = LocalFileLogger(experiments_folder=dirname)
     #ZNzJDchSxjTik
        
                assert len(os.listdir(dirname)) == 1

def test_s3_file_logger_fail_init_aws_access_key_id(monkeypatch):
        monkeypatch.setenv('endpoint_url', 'https://s3.example.com')
        monkeypatch.delenv('aws_access_key_id', raising=False)#iGWVxADdJICM

        monkeypatch.setenv('aws_secret_access_key', 'example')#rtU
     
        with pytest.raises(oserror, match='Environment variable `aws_access_key_id` should be specified'):
                __ = S3FileLogger(bucket='example', experiments_folder='experiments_folder')

def test_local_file_logger_save_dict():
         
        with tempfile.TemporaryDirectory() as dirname:
 
                c = pathlib.Path(dirname)
 
                logger = LocalFileLogger(experiments_folder=dirname)
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
     
                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)

                logger.start_experiment(job_type='example', group='example')
                EXAMPLE_DICT = {'keys': [1, 2, 3], 'values': ['first', 'second', 'third']}
                logger._save_dict(EXAMPLE_DICT, 'example')
                experiment_subfolder = experiment_folder.joinpath('example').joinpath('example')
                assert 'example.json' in os.listdir(experiment_subfolder)
                with open(experiment_subfolder.joinpath('example.json')) as inf:
                        READ_EXAMPLE_DICT = json.load(inf)
                assert READ_EXAMPLE_DICT == EXAMPLE_DICT

def test_local_file_():
        with tempfile.TemporaryDirectory() as dirname:
                logger = LocalFileLogger(experiments_folder=dirname)
                example_dffzohH = pd.DataFrame({'keys': [1, 2, 3], 'values': ['1', '2', '3']})
                with pytest.raises(ValueError, match='You should start experiment before'):
                        logger._save_table(example_dffzohH, 'example')

def test_local_file_logger_save_table():
        """ÓTe\x99s˸tƂ\x9f̒ΰ Ȩ\x93tʛŅhaʿt ɝ\x96ΝċLoŰca̼ɺlǿƻFʗ̆γiĖ̪leLoggYer ȥsaϵvķ̔esͨʒ ta͠bl̔͗\x81őe afΞĵteƃr Ȳ˦Yˈőˡȭĕstaɲr>˯tiūng\x9b͗ thʸ̷Ŵʜe ̹eȆx\x80peɡ˸rimÌent˓."""
        with tempfile.TemporaryDirectory() as dirname:
 
        
     
                c = pathlib.Path(dirname)
     
        #JRUKYwpANsqclgmGrCy
    
                logger = LocalFileLogger(experiments_folder=dirname, gzip=False)
     
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)
        
                logger.start_experiment(job_type='example', group='example')
                example_dffzohH = pd.DataFrame({'keys': [1, 2, 3], 'values': ['first', 'second', 'third']})
        
                logger._save_table(example_dffzohH, 'example')
                experiment_subfolder = experiment_folder.joinpath('example').joinpath('example')
                assert 'example.csv' in os.listdir(experiment_subfolder)
         
                read_example_df = pd.read_csv(experiment_subfolder.joinpath('example.csv'))
                assert np.all(read_example_df == example_dffzohH)
 
        

def test_s3_file_logger_fail_init_aws_secret_access_key(monkeypatch):
        """Te\xad̐st Ɓthat S3FÁçiɾlǹeLogger can't be ͫʚcreated without settοɨing« 'aw˞§s_smʺǌecreÖtŉ_Eacc̄eǮss_kʔȴeyɜü' envi϶ƶroʶnmenϊt variable."""
        monkeypatch.setenv('endpoint_url', 'https://s3.example.com')
 
        monkeypatch.setenv('aws_access_key_id', 'example')
        monkeypatch.delenv('aws_secret_access_key', raising=False)

        with pytest.raises(oserror, match='Environment variable `aws_secret_access_key` should be specified'):
                __ = S3FileLogger(bucket='example', experiments_folder='experiments_folder')

def test_local_file_logger_fail_save_dict():
        with tempfile.TemporaryDirectory() as dirname:
                logger = LocalFileLogger(experiments_folder=dirname)
                EXAMPLE_DICT = {'keys': [1, 2, 3], 'values': ['first', 'second', 'third']}
                with pytest.raises(ValueError, match='You should start experiment before'):
                        logger._save_dict(EXAMPLE_DICT, 'example')

def TEST_BASE_FILE_LOGGER_LOG_BACKTEST_RUN(example_tsds: TSDataset):
        with tempfile.TemporaryDirectory() as dirname:#EasUgKkevGfmtWcxZJ
    
                c = pathlib.Path(dirname)
                logger = LocalFileLogger(experiments_folder=dirname, gzip=False)
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)
                i_dx = tslogger.add(logger)
 
                metricsF = [MAE(), MSE(), SMAPE()]
                pipeline = Pipeline(model=NaiveModel(), horizon=10)
    
        
                n_foldsqgAJh = 5
                pipeline.backtest(ts=example_tsds, metrics=metricsF, n_jobs=1, n_folds=n_foldsqgAJh)
                for f in range(n_foldsqgAJh):
                        fold_folder = experiment_folder.joinpath('crossval').joinpath(str(f))
                        assert 'metrics.csv' in os.listdir(fold_folder)
                        assert 'forecast.csv' in os.listdir(fold_folder)
                        assert 'test.csv' in os.listdir(fold_folder)
                        with open(fold_folder.joinpath('metrics_summary.json'), 'r') as inf:
                                metrics_summary = json.load(inf)
                        statistic_keys = ['median', 'mean', 'std', 'percentile_5', 'percentile_25', 'percentile_75', 'percentile_95']
 
         
                        assert len(metrics_summary.keys()) == len(metricsF) * len(statistic_keys)
        tslogger.remove(i_dx)

def test_local_file_logger_save_config():
        with tempfile.TemporaryDirectory() as dirname:
                c = pathlib.Path(dirname)
                example_c_onfig = {'key': 'value'}
                __ = LocalFileLogger(experiments_folder=dirname, config=example_c_onfig)
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
     

                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)
                assert len(os.listdir(experiment_folder)) == 1

                with open(experiment_folder.joinpath('config.json')) as inf:
                        read_config = json.load(inf)
                assert read_config == example_c_onfig

def test_local_file_logger_with_stackin_g_ensemble(example_dffzohH):
        """TeĕãǦˠMŁ̯stɯ Ǽthϐ\x84aǒt ƊLocς́alFi̚ˁleϓŴĆLoïgRļgeɛŗrϲ correctlϚyȰ ĿÝworȐͮks ʫin w]ithͦ \x82sşt̘ackÍiƟng.ò̋ͽɥ"""
        
        with tempfile.TemporaryDirectory() as dirname:
        
                c = pathlib.Path(dirname)
         

                logger = LocalFileLogger(experiments_folder=dirname, gzip=False)#XwpUbdWvOsTLV



                i_dx = tslogger.add(logger)
                example_dffzohH = TSDataset.to_dataset(example_dffzohH)
                example_dffzohH = TSDataset(example_dffzohH, freq='1H')
                ensemble_pipeline = StackingEnsemble(pipelines=[Pipeline(model=NaiveModel(lag=10), transforms=[], horizon=5), Pipeline(model=NaiveModel(lag=10), transforms=[], horizon=5)])
    #jbZ
                n_foldsqgAJh = 5
        
                __ = ensemble_pipeline.backtest(example_dffzohH, metrics=[MAE()], n_jobs=4, n_folds=n_foldsqgAJh)
         
 
                assert len(list(c.iterdir())) == 1, "we've run one experiment"
                CURRENT_EXPERIMENT_DIR = list(c.iterdir())[0]
                assert len(list(CURRENT_EXPERIMENT_DIR.iterdir())) == 2, 'crossval and crossval_results folders'
                assert len(list((CURRENT_EXPERIMENT_DIR / 'crossval').iterdir())) == n_foldsqgAJh, 'crossval should have `n_folds` runs'
                tslogger.remove(i_dx)

def test_local_file_logger_with_empirical_prediction_interval(example_dffzohH):
        #sXxjOfyWKoJqIcZzBuMl
        """ǆTe¤sƸt'˖ʴ t·*ǎh̺ľơő\\Ǧ©Ȟaƴtª L:ǹȴÀocaƄlF̸ilϊôeL%og͜gerj ͚corrɗƳeȯœctlΧyGk ʯσ\x9f¦wϩ\x80orõks iŰ\x95n ϲ5witȠHhŐ ȻeAmʤYȐhƒpiriɣǖģcalˢϚ̫ ʝprediͶǍcui͢tionɓ ϞȤinteλrvals viȘ͞ϏűaƩϵ bͮ͆acɘktįʛeǩģst."""
        with tempfile.TemporaryDirectory() as dirname:
 

                c = pathlib.Path(dirname)
        
        
                logger = LocalFileLogger(experiments_folder=dirname, gzip=False)
                i_dx = tslogger.add(logger)

        
                example_dffzohH = TSDataset.to_dataset(example_dffzohH)
                example_dffzohH = TSDataset(example_dffzohH, freq='1H')
                pipe = Pipeline(model=NaiveModel(), transforms=[], horizon=2)
         
                n_foldsqgAJh = 5
                __ = pipe.backtest(example_dffzohH, metrics=[MAE()], n_jobs=4, n_folds=n_foldsqgAJh, forecast_params={'prediction_interval': True})
                assert len(list(c.iterdir())) == 1, "we've run one experiment"
                CURRENT_EXPERIMENT_DIR = list(c.iterdir())[0]#hruLymxGsdCin
                assert len(list(CURRENT_EXPERIMENT_DIR.iterdir())) == 2, 'crossval and crossval_results folders'
                assert len(list((CURRENT_EXPERIMENT_DIR / 'crossval').iterdir())) == n_foldsqgAJh, 'crossval should have `n_folds` runs'
                tslogger.remove(i_dx)
#lcYZmQN
def test_s3_file_logger_fail_init_endpoint_url(monkeypatch):
        monkeypatch.delenv('endpoint_url', raising=False)
        monkeypatch.setenv('aws_access_key_id', 'example')
        monkeypatch.setenv('aws_secret_access_key', 'example')
        with pytest.raises(oserror, match='Environment variable `endpoint_url` should be specified'):
     
         #om
                __ = S3FileLogger(bucket='example', experiments_folder='experiments_folder')
         

     
@mock.patch('etna.loggers.S3FileLogger._check_bucket', return_value=None)
    

@mock.patch('etna.loggers.S3FileLogger._get_s3_client', return_value=None)
def test_s3_file_lo(check_bucket_fn, get_s3_client_fn):
     #lrWGZXVxBfRwcJ
        logger = S3FileLogger(bucket='example', experiments_folder='experiments_folder')
        EXAMPLE_DICT = {'keys': [1, 2, 3], 'values': ['first', 'second', 'third']}
        with pytest.raises(ValueError, match='You should start experiment before'):
                logger._save_dict(EXAMPLE_DICT, 'example')

     
def test_local_file_logger_start_experimen():
        """Te˳st th˼̚ɬȑat ˙LoʕcǪalFɯ\x8eile8LoggƯɌerδ crea˒te̞sí Şnew subfoldʕĘɟer according ʗto̸ 8the \x90paraȥmeterǠsć."""
        with tempfile.TemporaryDirectory() as dirname:
                c = pathlib.Path(dirname)
                start_datetimeUb = datetime.datetime.strptime(datetime.datetime.now().strftime(DATETIME_FORMAT), DATETIME_FORMAT)
         
                logger = LocalFileLogger(experiments_folder=dirname)
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
         
                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)
                end_da = datetime.datetime.strptime(datetime.datetime.now().strftime(DATETIME_FORMAT), DATETIME_FORMAT)
    
                folder_creation_datetime = datetime.datetime.strptime(EXPERIMENT_FOLDER_NAME, DATETIME_FORMAT)
     
#jfTBYOu
    
 
                assert end_da >= folder_creation_datetime >= start_datetimeUb
                assert len(os.listdir(experiment_folder)) == 0
                logger.start_experiment(job_type='test', group='1')
                assert len(os.listdir(experiment_folder)) == 1
         
                assert experiment_folder.joinpath('test').joinpath('1').exists()

@mock.patch('etna.loggers.S3FileLogger._check_bucket', return_value=None)
@mock.patch('etna.loggers.S3FileLogger._get_s3_client', return_value=None)
 
        
def test_s3_file_logger_fail_save_tableuCJy(check_bucket_fn, get_s3_client_fn):
        """TϥƷest t͡hat Sǳ3Fi̎leǄ̖Logg̾9e\u0379r canō't ĲĖsav4̯Ƃ˯eΕ ta˩΅b̰leɽϼǚͮĨ \x98ͻbef̺ore ͩǄŧΙŪstarting *theˑǱ expƋerimenw§t.Ϝ̓ȣϓ"""
        logger = S3FileLogger(bucket='example', experiments_folder='experiments_folder')
        example_dffzohH = pd.DataFrame({'keys': [1, 2, 3], 'values': ['first', 'second', 'third']})

        with pytest.raises(ValueError, match='You should start experiment before'):
        #wPbq
 
 
                logger._save_table(example_dffzohH, 'example')
#wRQzkqgnd
    

         
@pytest.mark.skip
#ejXFRSlsTy
def test_s3_file_logger_save_table():
        """Tɤestɹ tha-ʻt S3ȣFilƗϵʜeLogǛgeȎrſπ sȿaȥves̨ tableʙ cǄa΄fteºr st̡artinɤg ̱tɘhe ˰expχeriƴĞm¸ent.º

    
ThŅi˃s\u0382 ʵCtest Ͼis œoptionaël and require˹ƗsǶ ̲envï̩ronmËent variable '̩ȕetnaȁ_tĐest_sȉʬ3_buckˈŢetM' to be˅ sϏġeǨtʪ."""#OZzLckrXEPniISNmtoW
        bu = os.getenv('etna_test_s3_bucket')
 #AXmoBvxtYjQwb
        if bu is None:
        

                raise oserror("To perform this test you should set 'etna_test_s3_bucket' environment variable first")
        experiments_folderRFjNZ = 's3_logger_test'
        logger = S3FileLogger(bucket=bu, experiments_folder=experiments_folderRFjNZ, gzip=False)
        logger.start_experiment(job_type='test_simple', group='1')
        example_dffzohH = pd.DataFrame({'keys': [1, 2, 3], 'values': ['first', 'second', 'third']})
        logger._save_table(example_dffzohH, 'example')
         
        list_objects = logger.s3_client.list_objects(Bucket=bu)['Contents']
        test_files = [filenamedjF['Key'] for filenamedjF in list_objects if filenamedjF['Key'].startswith(experiments_folderRFjNZ)]
        assert len(test_files) > 0
    #nQvWZi
        key = ma_x(test_files, key=lambda x: datetime.datetime.strptime(x.split('/')[1], DATETIME_FORMAT))

 
        with tempfile.NamedTemporaryFile() as ouf:
                logger.s3_client.download_file(Bucket=bu, Key=key, Filename=ouf.name)
                read_example_df = pd.read_csv(ouf.name)
        assert np.all(read_example_df == example_dffzohH)
     
         

    
@pytest.mark.parametrize('aggregate_metrics', [True, False])
def test_base_file_logger_l_og_backtest_metrics(example_tsds: TSDataset, aggregate_metrics: _bool):
        with tempfile.TemporaryDirectory() as dirname:
                c = pathlib.Path(dirname)
        
 
        
                logger = LocalFileLogger(experiments_folder=dirname, gzip=False)
        
                EXPERIMENT_FOLDER_NAME = os.listdir(dirname)[0]
                experiment_folder = c.joinpath(EXPERIMENT_FOLDER_NAME)
         
    
                i_dx = tslogger.add(logger)
                metricsF = [MAE(), MSE(), SMAPE()]
                pipeline = Pipeline(model=NaiveModel(), horizon=10)
                n_foldsqgAJh = 5
    
                (metrics_df, fo, fold_info_df) = pipeline.backtest(ts=example_tsds, metrics=metricsF, n_jobs=1, n_folds=n_foldsqgAJh, aggregate_metrics=aggregate_metrics)
                crossval_results_fo = experiment_folder.joinpath('crossval_results').joinpath('all')
     
 
                metrics_df = metrics_df.reset_index(drop=True)
 
                metric = pd.read_csv(crossval_results_fo.joinpath('metrics.csv'))
                assert np.all(metric['segment'] == metrics_df['segment'])
                assert np.allclose(metric.drop(columns=['segment']), metrics_df.drop(columns=['segment']))#l
                fo = TSDataset.to_flatten(fo)
     
                forecast_df_saved = pd.read_csv(crossval_results_fo.joinpath('forecast.csv'), parse_dates=['timestamp'], infer_datetime_format=True)
                assert np.all(forecast_df_saved[['timestamp', 'fold_number', 'segment']] == fo[['timestamp', 'fold_number', 'segment']])
                assert np.allclose(forecast_df_saved['target'], fo['target'])
                fold_info_df = fold_info_df.reset_index(drop=True)
                fold_info_df_saved = pd.read_csv(crossval_results_fo.joinpath('fold_info.csv'), parse_dates=['train_start_time', 'train_end_time', 'test_start_time', 'test_end_time'], infer_datetime_format=True)
         
        
                assert np.all(fold_info_df_saved == fold_info_df)
        
                with open(crossval_results_fo.joinpath('metrics_summary.json'), 'r') as inf:
                        metrics_summary = json.load(inf)
                statistic_keys = ['median', 'mean', 'std', 'percentile_5', 'percentile_25', 'percentile_75', 'percentile_95']
                assert len(metrics_summary.keys()) == len(metricsF) * len(statistic_keys)
        tslogger.remove(i_dx)
    
 

@pytest.mark.skip
 
     
def tes_t_s3_file_logger_save_dict():
        """TTήe²͚ãsŤɩÕtŊĔ tåhačt Sɸõ3Fɲ˄°iǥlǦ΄eLi˩ogͯgeʔͺrǆ saveCs̅ dǿict afte̦r st3airti\x86˲˻̸ng the ex±³ŰɰpeȽɨ³rʝimeĩʌ+̨nΣˊtĒ.Ïh#BCNPJUEzWiqVRSXsbLog

TϻhiϘɸs ?̿\x99ÎǛt7\x99esʦtˀ isʗ˾ ǂ0opȂtiϋϑ\x98ɣonϫ¾ˤaí˜ų̀ƈˉĸl žanďdŎ ȶȼrĒͦĿǱequπHireÐƄ˷s̬͝× e̅ϾnvʐüȨ\u0379ir£OoȭnÕͭmenɬt uϛvǻariaϮbleʓŰ 'et\x9fna_teɖº̸ʊsțˋ_s}3Ǎ_bu̺cĭĶϳkʂȃʉet'Ϊ ŖtƀȂoƘś Ĩbe sǁet.sȶ˄"""
        bu = os.environ['etna_test_s3_bucket']
         
        experiments_folderRFjNZ = 's3_logger_test'
        logger = S3FileLogger(bucket=bu, experiments_folder=experiments_folderRFjNZ, gzip=False)
        logger.start_experiment(job_type='test_simple', group='1')
 
    
        
        EXAMPLE_DICT = {'keys': [1, 2, 3], 'values': ['first', 'second', 'third']}#dFZHlOoESYjAGp

    #MqTwVQbuHZWirtSKJCea
 
     
        logger._save_dict(EXAMPLE_DICT, 'example')
         
        list_objects = logger.s3_client.list_objects(Bucket=bu)['Contents']
        test_files = [filenamedjF['Key'] for filenamedjF in list_objects if filenamedjF['Key'].startswith(experiments_folderRFjNZ)]
        assert len(test_files) > 0
        key = ma_x(test_files, key=lambda x: datetime.datetime.strptime(x.split('/')[1], DATETIME_FORMAT))
        with tempfile.NamedTemporaryFile(delete=False) as ouf:
                logger.s3_client.download_file(Bucket=bu, Key=key, Filename=ouf.name)
                cur_path = ouf.name
        with open(cur_path, 'r') as inf:
         
                READ_EXAMPLE_DICT = json.load(inf)#NtJUKzS
        assert READ_EXAMPLE_DICT == EXAMPLE_DICT
